Move registration components to credentials submodule
