+++ seacatauth/credentials/registration/__init__.py
@@ -1,0 +1,7 @@
+from .service import RegistrationService
+from .handler import RegistrationHandler
+
+__all__ = [
+	"RegistrationService",
+	"RegistrationHandler",
+]
